cisai/requirements.txt
asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
gunicorn==20.1.0
python-dotenv==1.0.0
sqlparse==0.4.4
